+++ modules/conversion-center/widgets/link-in-bio.php
@@ -28,44 +28,4 @@ class Link_In_Bio extends Widget_Link_In_Bio_Base {
 		return esc_html__( 'Link In Bio', 'elementor' );
 	}
 
-	public function get_icon(): string {
-		return 'eicon-bullet-list';
-	}
-
-	public function get_categories(): array {
-		return [ 'general' ];
-	}
-
-	public function get_keywords(): array {
-		return [ 'buttons', 'bio', 'widget' ];
-	}
-
-	public function show_in_panel(): bool {
-		return Plugin::$instance->experiments->is_feature_active( ConversionCenterModule::EXPERIMENT_NAME );
-	}
-
-	public function get_stack( $with_common_controls = true ): array {
-		return parent::get_stack( false );
-	}
-
-	protected function register_controls(): void {
-
-		$this->add_identity_section();
-
-		$this->add_bio_section();
-
-		$this->add_icons_controls();
-
-		$this->add_cta_controls();
-
-		$this->add_style_tab();
-
-		$this->add_advanced_tab();
-	}
-
-	protected function render(): void {
-		$render_strategy = new Core_Render( $this );
-
-		$render_strategy->render();
-	}
 }
